Add Log4j2 appender autoconfiguration with multi-module structure
Convert single-module project to multi-module layout to support both
Logback and Log4j2 appenders without unnecessary transitive dependencies.

- otel-logs-autoconfigure-logback: existing Logback implementation (moved)
- otel-logs-autoconfigure-log4j: new Log4j2 appender autoconfiguration
- otel-logs-autoconfigure (legacy): backward-compatible wrapper for existing users
