+++ otel-logs-autoconfigure-log4j/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.context.ApplicationListener=\
+am.ik.spring.opentelemetry.logs.log4j.Log4jAppenderInstallListener
