--- src/main/java/am/ik/spring/opentelemetry/logs/logback/LogbackAppenderProps.java
@@ -1,8 +1,0 @@
-package am.ik.spring.opentelemetry.logs.logback;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "management.opentelemetry.instrumentation.logback-appender")
-public class LogbackAppenderProps {
-
-}
